feature bitfix
Move struct to tweet.go
Add the reason why it is not supported
